requirements.txt
ttkbootstrap
pynput
python-xlib
